backend/requirements.txt
alabaster==0.7.13
asgiref==3.6.0
Babel==2.13.1
boto3==1.26.123
botocore==1.29.123
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Django==4.1.9
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
docutils==0.20.1
drf-yasg==1.21.7
idna==3.6
imagesize==1.4.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.2
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.17.2
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
simplejson==3.19.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.18
gunicorn==21.2.0
google-cloud-storage==2.14.0